// Unmatched.UI.Angular/src/app/backend-api/models/ui-match-log-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UiFighterDto } from '../models/ui-fighter-dto';
export interface UiMatchLogDto {
  comment?: string | null;
  date?: string;
  epic?: number | null;
  fighters?: Array<UiFighterDto> | null;
  mapName?: string | null;
  matchId?: string;
  tournamentName?: string | null;
}